refactor(shared): split project host setup lookup
